ISSUE #1300: Moved classes to proper namespaces
