--- src/Domain/Dashboard/WeeklyGoals/WeeklyGoalType.php
@@ -1,20 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Domain\Dashboard\WeeklyGoals;
-
-enum WeeklyGoalType: string
-{
-    case DISTANCE = 'distance';
-    case ELEVATION = 'elevation';
-    case MOVING_TIME = 'movingTime';
-
-    /**
-     * @return WeeklyGoalType[]
-     */
-    public static function lengthRelated(): array
-    {
-        return [self::DISTANCE, self::ELEVATION];
-    }
-}
