+++ templates/html/calendar/monthly-stats.html.twig
@@ -7,12 +7,6 @@
     <!-- MONTHLY STATS -->
     <div class="grid grid-cols-1 mb-4">
         <div class="p-4 bg-white border border-gray-200 rounded-lg shadow-xs">
-            <div class="flex items-center justify-end mb-2">
-                <a href="#" data-model-content-url="{{ relativeUrl('monthly-stats/charts.html') }}"
-                   class="inline-flex items-center p-2 text-sm font-medium rounded-lg text-blue-600 hover:bg-gray-100">
-                    {{ "View details"|trans }}
-                </a>
-            </div>
             <div class="overflow-auto max-h-[calc(100vh-200px)] relative">
                 <table class="w-full text-sm text-center text-gray-500" data-accordion="table-months"
                        data-active-classes="text-gray-500" data-inactive-classes="text-gray-500">
